Revert "move totals into subfolder"
This reverts commit 4e6e29e0
