+++ libs/domain/cart/totals/components/delivery/stories/static.stories.ts
