+++ libs/domain/cart/totals/components/discount/discount.component.spec.ts
