+++ libs/domain/cart/totals/components/discount/discount.component.ts
@@ -2,8 +2,8 @@ import { ContentMixin, defaultOptions } from '@spryker-oryx/experience';
 import { CollapsibleAppearance } from '@spryker-oryx/ui/collapsible';
 import { hydrate, signal, signalAware } from '@spryker-oryx/utilities';
 import { LitElement, TemplateResult, html } from 'lit';
-import { TotalsController } from '../../../../src/controllers';
-import { CartDiscount } from '../../../../src/models';
+import { TotalsController } from '../../../src/controllers';
+import { CartDiscount } from '../../../src/models';
 import {
   CartTotalsDiscountOptions,
   DiscountRowsAppearance,
