+++ libs/domain/cart/totals/components/discount/discount.styles.ts
