+++ libs/domain/cart/totals/components/discount/stories/static.stories.ts
