+++ libs/domain/cart/totals/components/expense/expense.component.ts
@@ -5,22 +5,22 @@ import {
   signalAware,
 } from '@spryker-oryx/utilities';
 import { html, LitElement, TemplateResult } from 'lit';
-import { TotalsController } from '../../../../src/controllers';
+import { TotalsController } from '../../../src/controllers';
 
 @hydrate({ event: 'window:load' })
 @signalAware()
-export class CartTotalsTaxComponent extends I18nMixin(LitElement) {
+export class CartTotalsExpenseComponent extends I18nMixin(LitElement) {
   protected totalsController = new TotalsController(this);
 
   protected $totals = signal(this.totalsController.getTotals());
 
   protected override render(): TemplateResult | void {
-    const { taxTotal, currency } = this.$totals() ?? {};
-    if (taxTotal) {
+    const { expenseTotal, currency } = this.$totals() ?? {};
+    if (expenseTotal) {
       return html`
-        <span>${this.i18n('cart.totals.tax')}</span>
+        <span>${this.i18n('cart.totals.expense')}</span>
         <oryx-site-price
-          .value=${taxTotal}
+          .value=${expenseTotal}
           .currency=${currency}
         ></oryx-site-price>
       `;
