+++ libs/domain/cart/totals/components/subtotal/subtotal.component.ts
@@ -5,22 +5,22 @@ import {
   signalAware,
 } from '@spryker-oryx/utilities';
 import { html, LitElement, TemplateResult } from 'lit';
-import { TotalsController } from '../../../../src/controllers';
+import { TotalsController } from '../../../src/controllers';
 
 @hydrate({ event: 'window:load' })
 @signalAware()
-export class CartTotalsExpenseComponent extends I18nMixin(LitElement) {
+export class CartTotalsSubtotalComponent extends I18nMixin(LitElement) {
   protected totalsController = new TotalsController(this);
 
   protected $totals = signal(this.totalsController.getTotals());
 
   protected override render(): TemplateResult | void {
-    const { expenseTotal, currency } = this.$totals() ?? {};
-    if (expenseTotal) {
+    const { subtotal, currency } = this.$totals() ?? {};
+    if (subtotal) {
       return html`
-        <span>${this.i18n('cart.totals.expense')}</span>
+        <span>${this.i18n('cart.totals.subtotal')}</span>
         <oryx-site-price
-          .value=${expenseTotal}
+          .value=${subtotal}
           .currency=${currency}
         ></oryx-site-price>
       `;
