+++ libs/domain/cart/totals/components/total/stories/static.stories.ts
