+++ libs/domain/cart/totals/components/total/total.component.ts
@@ -1,8 +1,8 @@
 import { ContentMixin, defaultOptions } from '@spryker-oryx/experience';
 import { hydrate, signal, signalAware } from '@spryker-oryx/utilities';
 import { LitElement, TemplateResult, html } from 'lit';
-import { TotalsController } from '../../../../src/controllers';
-import { PriceMode } from '../../../../src/models';
+import { TotalsController } from '../../../src/controllers';
+import { PriceMode } from '../../../src/models';
 import { CartTotalsTotalOptions } from './total.model';
 
 @defaultOptions({ enableTaxMessage: true })
